Polish and Repair Programmatic Unit Tests
This commit shuts off Log4j2 annotation processing of custom plugins
so that they can be declaratively included in the relevant unit
tests.

Because Log4j2 tracks plugins statically though, it is necessary to
clean up this state after each test that configures them. To that
end, this commit also introduces a Log4j2Test parent class with
an @AfterClass that attempts to do this cleanup.

Also, the test names are now according to the Baeldung naming
convention.

Further, I did not see a reason to keep the two xml tests separate,
and so I have combined them in the hopes of simplifying things for
the user.

Finally, note that there is still a discrepancy in the way these
tests run on the command line vs the IDE, but I have not yet been
able to track it down. All the tests run separately just fine and
on the command line, but run them together in the IDE and there are
some failures.

Issue: BAEL-1338