BAEL-1338 - Programmatic Log4j2
This commit features a number of tests for programmatically configuring log4j.

Note that in order to have these tests live together, there is a slight deviation in how
these are configured vs how they are in the real world, specifically around static initialization.

Generally speaking, whenever static initialization is needed for a given configuration,
it is embedded in a @BeforeClass-annotated JUnit method.

Also, to isolate the tests from each other's plugins, annotation processing was turned off
in the build. This is not required when doing programmatic configuration of log4j. It is simply
needed for the specific use case of demonstrating these various log4j plugins all in the
same module.