Update PostRestController.java
The id is already part of the postDto in this case. And the controller method does not have a @PathVariable annotation to evaluate the /{id} parameter. Guess the id should be transmitted in the RequestBody in the PostDto Object? Otherwise, id would be redundant, once in the PathVariable and once in the PostDto.